requirements.txt
altgraph==0.17.4
colorama==0.4.6
docopt==0.6.2
et_xmlfile==2.0.0
lxml==6.0.0
num2words==0.5.14
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pefile==2023.2.7
pillow==11.3.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pywin32-ctypes==0.2.3
six==1.17.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
